adds sourcefiles as dependency to build task
Monitoring the modified state of source files for the build task means
that simply running 'make' will detect if a rebuild is actually needed
or not. Currently if the build file exists make is not run again. A
clean must be run first, then make. This is an improvement on that.