Fixes a 'resource not found' bug
When navigating to many sites, such as gemini://rawtext.club:1965/~sloum/geminilist and not having a trailing slash at the end of the URL, local links on the page would be incorrectly generated by QUrl::resolved, which would cause links to give a 'resource not found' error if they weren't at the root path of the server. This patch fixes this by making sure that all links given to the renderers include trailing slashes.