Fix window geometry/state not being saved in some environments
The window state and geometry are now saved either after app.exec() finished, or if closeEvent is emitted on the MainWindow.

This allows the window geometry/state restoration to be a lot more reliable. E.g in my configuration (Fluxbox) - if I closed the window using my own defined keybinding, the state would not get saved