KristallTextBrowser: replace typographer quotes with ASCII when copying text
Text which is selected in a kristalltextbrowser now has it's quotes replaced with ASCII quotes. This was done as copying fancy quotes is seen as an annoyance to most people. It is still possible to copy these quotes, if they are the only thing in the selection (whitespace also allowed, and commas, and fullstops)