Makefile: remove MAKEFLAGS parameter from sub-make calls
This broke running make with flags like -k or -s set,
because make apparently removes hyphens from those flags,
making it to think that we want to run a not existing target.

Also, flags are stored in env, so child processes know which flags
are in use and we don't have to provide them again.