Unwarp tr() from QStrings
We don't need to wrap tr() with QString, because it returns QString anyway.
Although it doesn't change the binary size (mostly because compilers were
smart to remove that already), it helps to make the code a little bit cleaner.

It's a somewhat continuation of c740189bcaa5bc6c073f8b3802303acf5ac36e47,
but I thought back then it happened only there, so now I've removed it from
all places.